requirements.txt
azure.core
azure-ai-documentintelligence
azure-storage-blob
streamlit
python-dotenv